megano/requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.2.6
django-filter==23.3
djangorestframework==3.14.0
drf-spectacular==0.26.5
inflection==0.5.1
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
Pillow==10.1.0
pip==23.3.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.10.6
setuptools==69.0.2
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
wheel==0.42.0
python-dotenv
gunicorn==21.2.0
psycopg2-binary==2.9.6
celery[redis]==5.2.7
flower==1.2.0
celery_singleton==0.3.1
django-redis==5.2.0
#psycopg2==2.8.6
